Initial couchbase integration completed; refactored to reduce redundancy
